+++ ntex/tests/http_server.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 use std::{io, net, thread};
 
 use bytes::Bytes;
-use futures::future::{self, err, ok, ready, FutureExt};
+use futures::future::{self, ok, ready, FutureExt};
 use futures::stream::{once, StreamExt};
 use regex::Regex;
 
@@ -60,52 +60,21 @@ async fn test_h1_2() {
 async fn test_expect_continue() {
     let srv = test_server(|| {
         HttpService::build()
-            .expect(fn_service(|req: Request| {
+            .expect(fn_service(|req: Request| async move {
+                sleep(Duration::from_millis(20)).await;
                 if req.head().uri.query() == Some("yes=") {
-                    ok(req)
+                    Ok(req)
                 } else {
-                    err(error::InternalError::default(
+                    Err(error::InternalError::default(
                         "error",
                         StatusCode::PRECONDITION_FAILED,
                     ))
                 }
             }))
-            .finish(|_| future::ok::<_, io::Error>(Response::Ok().finish()))
-            .tcp()
-    });
-
-    let mut stream = net::TcpStream::connect(srv.addr()).unwrap();
-    let _ = stream.write_all(b"GET /test HTTP/1.1\r\nexpect: 100-continue\r\n\r\n");
-    let mut data = String::new();
-    let _ = stream.read_to_string(&mut data);
-    assert!(data.starts_with("HTTP/1.1 412 Precondition Failed\r\ncontent-length"));
-
-    let mut stream = net::TcpStream::connect(srv.addr()).unwrap();
-    let _ = stream.write_all(b"GET /test?yes= HTTP/1.1\r\nexpect: 100-continue\r\n\r\n");
-    let mut data = String::new();
-    let _ = stream.read_to_string(&mut data);
-    assert!(data.starts_with("HTTP/1.1 100 Continue\r\n\r\nHTTP/1.1 200 OK\r\n"));
-}
-
-#[ntex::test]
-async fn test_expect_continue_h1() {
-    let srv = test_server(|| {
-        HttpService::build()
-            .expect(fn_service(|req: Request| {
-                sleep(Duration::from_millis(20)).then(move |_| {
-                    if req.head().uri.query() == Some("yes=") {
-                        ok(req)
-                    } else {
-                        err(error::InternalError::default(
-                            "error",
-                            StatusCode::PRECONDITION_FAILED,
-                        ))
-                    }
-                })
-            }))
             .keep_alive(KeepAlive::Disabled)
-            .h1(fn_service(|_| {
-                future::ok::<_, io::Error>(Response::Ok().finish())
+            .h1(fn_service(|mut req: Request| async move {
+                let _ = req.payload().next().await;
+                Ok::<_, io::Error>(Response::Ok().finish())
             }))
             .tcp()
     });
@@ -117,10 +86,17 @@ async fn test_expect_continue_h1() {
     assert!(data.starts_with("HTTP/1.1 412 Precondition Failed\r\ncontent-length"));
 
     let mut stream = net::TcpStream::connect(srv.addr()).unwrap();
-    let _ = stream.write_all(b"GET /test?yes= HTTP/1.1\r\nexpect: 100-continue\r\n\r\n");
+    let _ = stream.write_all(
+        b"GET /test?yes= HTTP/1.1\r\ncontent-length:4\r\nexpect: 100-continue\r\n\r\n",
+    );
+    let mut data = [0; 25];
+    let _ = stream.read_exact(&mut data[..]);
+    assert_eq!(&data, b"HTTP/1.1 100 Continue\r\n\r\n");
+
     let mut data = String::new();
+    let _ = stream.write_all(b"test");
     let _ = stream.read_to_string(&mut data);
-    assert!(data.starts_with("HTTP/1.1 100 Continue\r\n\r\nHTTP/1.1 200 OK\r\n"));
+    assert!(data.starts_with("HTTP/1.1 200 OK\r\n"));
 }
 
 #[ntex::test]